main_app/requirements.txt
aiohttp==3.8.5
fastapi==0.100.1
Pillow==9.5.0
pydantic==2.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
streamlit==1.25.0
uvicorn==0.23.2
opencv-python
